+++ src/main/java/codesquad/codestagram/config/SpringConfig.java
@@ -1,6 +1,8 @@
 package codesquad.codestagram.config;
 
-import codesquad.codestagram.user.repository.MemoryUserRepository;
+import codesquad.codestagram.article.repository.ArticleRepository;
+import codesquad.codestagram.article.repository.impl.MemoryArticleRepository;
+import codesquad.codestagram.user.repository.impl.MemoryUserRepository;
 import codesquad.codestagram.user.repository.UserRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
