+++ src/main/resources/templates/index.html
@@ -66,7 +66,7 @@
                     </ul>
                 </div>
                 <div class="col-md-3 qna-write">
-                    <a href="./qna/form.html" class="btn btn-primary pull-right" role="button">질문하기</a>
+                    <a th:href="@{/articles/form}" href="./qna/form.html" class="btn btn-primary pull-right" role="button">질문하기</a>
                 </div>
             </div>
         </div>
